dom/file/tests/test_blob_reading.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Reading blobs</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="common_blob_reading.js"></script>
  <script src="common_blob_types.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <script>

SimpleTest.waitForExplicitFinish();

async function runAllTests() {
  let content = "hello world";
  await forEachBlobType(content, async blob => {
    await testBlobText(blob, content);
    await workify('testBlobText', blob, content);

    await testBlobArrayBuffer(blob, content);
    await workify('testBlobArrayBuffer', blob, content);

    await testBlobStream(blob, content);
    await workify('testBlobStream', blob, content);
  });
}

runAllTests().then(SimpleTest.finish);

  </script>
</body>
</html>
